AddDialog devided on two parts for correct unmounting after dialog is closed
